requirements.txt
fastapi
mysql-connector-python
uvicorn[standard]
python-dotenv